requirements.txt
fastapi
uvicorn
yfinance
pandas
truedata-ws
python-dotenv
requests
